V2: move classes to fused.py
Differential Revision: D16349133
Differential Version: 86729609
