+++ torch/nn/_intrinsic/modules/__init__.py
@@ -1,21 +1,5 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
-import torch
-
-class ConvReLU2d(torch.nn.Sequential):
-    def __init__(self, conv, relu):
-        super(ConvReLU2d, self).__init__(conv, relu)
-
-class LinearReLU(torch.nn.Sequential):
-    def __init__(self, linear, relu):
-        super(LinearReLU, self).__init__(linear, relu)
-
-class ConvBn2d(torch.nn.Sequential):
-    def __init__(self, conv, bn):
-        super(ConvBn2d, self).__init__(conv, bn)
-
-class ConvBnReLU2d(torch.nn.Sequential):
-    def __init__(self, conv, bn, relu):
-        super(ConvBnReLU2d, self).__init__(linear, bn, relu)
+from .module.fused import ConvReLU2d, LinearReLU, ConvBn2d, ConvBnReLU2d
 
 
 __all__ = [
